update AES Encrypt Test

+++ tests/seismic/sol/AESEncrypt.sol
@@ -8,7 +8,7 @@ contract AESENCRYPT {
     /// @param plaintext The bytes to encrypt.
     /// @param nonce the u64 nonce for encryption, encoded as a big-endian bytes32.
     /// returns the encrypted bytes.
-    function AESEncrypt(bytes32 aes_key, uint256 nonce, bytes memory plaintext) public view returns (bytes memory) {
+    function AESEncrypt(bytes32 aes_key, uint96 nonce, bytes memory plaintext) public view returns (bytes memory) {
         // Address of the precompiled contract
         address AESEncryptAddr = address(0x67);
 
@@ -26,7 +26,7 @@ contract AESENCRYPT {
     }
 
     function testAESEncrypt() public view returns (bytes memory result) {
-        uint256 nonce = 17;
+        uint96 nonce = 17;
         bytes32 aes_key = hex"00112233445566778899aabbccddeeff00112233445566778899aabbccddeeff";
         bytes memory plaintext = hex"0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f202122232425262728292a2b2c2d2e2f303132333435363738393a3b3c3d3e3f";
         result = AESEncrypt(aes_key, nonce, plaintext);
@@ -37,4 +37,4 @@ contract AESENCRYPT {
 // EVMVersion: >=mercury
 // ====
 // ----
-// testAESEncrypt() -> hex"0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000004ff577b2b34b7dbafad6647accfaa9194d7a39c839e618fdbe9fc304691385c6fdcb1a8bf1c84560871726c31334884d85b463b0d9930c50370b9cdcc492dfcfb232dd38f0b0beb1c75e6f5c07e3a9ad0000000000000000000000000000000000"
+// testAESEncrypt() -> hex"00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000053cbf232bfd91fe00594508e3615f2f8a51f3de3d5965837cd25b550b38be2a8d5b356328a0cf91e892d0177d3a3c2ceebe5334a1655353f07d5ca1c98b442e35e39b3c86327fce9a76fe77a331bfcb510aed78200000000000000000000000000"
